(refactor): Package refactor
Package 'messages' has been deleted. Instead peeps and bulletin are in
the corresponding packages
