+++ src/main/java/acme/entities/bulletin/BulletinMessage.java
@@ -1,12 +1,11 @@
 
-package acme.entities.messages;
+package acme.entities.bulletin;
 
 import java.util.Date;
 
 import javax.persistence.Entity;
 import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
-import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PastOrPresent;
@@ -22,7 +21,7 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-public class PeepMessage extends AbstractEntity {
+public class BulletinMessage extends AbstractEntity {
 
 	/**
 	 * 
@@ -46,11 +45,6 @@ public class PeepMessage extends AbstractEntity {
 	@URL
 	protected String			link;
 
-	@Email
-	protected String			email;
-
-	@NotBlank
-	@Size(max = 75)
-	protected String			nickname;
+	protected boolean			isCritical;
 
 }
