+++ src/main/java/acme/features/any/peep/PeepMessageShowService.java
@@ -1,20 +1,16 @@
 
-package acme.features.messages.peep;
-
-import java.util.Collection;
+package acme.features.any.peep;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import acme.entities.messages.PeepMessage;
+import acme.entities.peep.PeepMessage;
 import acme.framework.components.accounts.Any;
 import acme.framework.components.models.Tuple;
 import acme.framework.services.AbstractService;
 
 @Service
-public class PeepMessageListService extends AbstractService<Any, PeepMessage> {
-
-	// Internal state ---------------------------------------------------------
+public class PeepMessageShowService extends AbstractService<Any, PeepMessage> {
 
 	@Autowired
 	protected PeepMessageRepository repository;
@@ -29,21 +25,33 @@ public class PeepMessageListService extends AbstractService<Any, PeepMessage> {
 
 	@Override
 	public void authorise() {
-		super.getResponse().setAuthorised(true);
+		boolean status;
+		int id;
+		PeepMessage peep;
+
+		id = super.getRequest().getData("id", int.class);
+		peep = this.repository.findOnePeepById(id);
+		status = peep != null;
+
+		super.getResponse().setAuthorised(status);
 	}
 
 	@Override
 	public void load() {
-		Collection<PeepMessage> peeps;
-		peeps = this.repository.findAllPeeps();
-		super.getBuffer().setData(peeps);
+		PeepMessage peep;
+		int id;
+
+		id = super.getRequest().getData("id", int.class);
+		peep = this.repository.findOnePeepById(id);
+
+		super.getBuffer().setData(peep);
 	}
 
 	@Override
 	public void unbind(final PeepMessage peep) {
 		assert peep != null;
 		Tuple tuple;
-		tuple = super.unbind(peep, "title", "instantiation", "nickname", "link", "email", "message");
+		tuple = super.unbind(peep, "title", "message", "instantiation", "nickname", "email", "link");
 		super.getResponse().setData(tuple);
 	}
