+++ src/main/java/acme/features/authenticated/bulletin/BulletinMessageCreateService.java
@@ -1,10 +1,10 @@
 
-package acme.features.messages.bulletin;
+package acme.features.authenticated.bulletin;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import acme.entities.messages.BulletinMessage;
+import acme.entities.bulletin.BulletinMessage;
 import acme.framework.components.accounts.Authenticated;
 import acme.framework.components.models.Dataset;
 import acme.framework.components.models.Tuple;
