+++ src/test/java/acme/testing/peep/PeepTestRepository.java
@@ -6,7 +6,7 @@ import java.util.Collection;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import acme.entities.messages.PeepMessage;
+import acme.entities.peep.PeepMessage;
 import acme.framework.repositories.AbstractRepository;
 
 @Repository
